Added optional pathfinder and initializer
The pathfinder function is now passed via the config hash in
resource-api and is used to determine the resulting $path of a
resource after a POST request has completed.

The initializer function is passed via the config hash in resource-api
and is invoked on a resource immediately after the data promise is
resolved and before the $loaded flag is set.