Factory for collections of resources
It is frequently useful to maintain similar resources in a specialized
collection. Collections are sort of resources in their own right in that
they map to specific paths and can be populated asynchronously with a
promise. Collections maintain a mapped array of resource objects as well
as methods to interact with those resources.