Ability to forcefully reload cached resources
Any call to .get() that includes a reload flag will always fire off the
corresponding HTTP request and return the stored resource. When the HTTP
request is completed, the stored resource is updated with the response
body.

If the resource was already stored, the only effect is to asynchronously
set the resource's data, so no initializer function will be reapplied to
the resource, and the $promise will not change.