Ability to forcefully reload cached collections
Any call to .query() that includes a reload flag will always fire off
the corresponding HTTP request and return the stored collection. When
the HTTP request is completed, the stored collection is updated with the
response body.

If the collection was already stored, then any resources not in the http
response are removed from the collection. Any existing resources in the
http response are updated (via .$extend()) in the collection/cache. Any
resources in the http response that were not already in the
collection/cache are initialized with the given config.intializer and
added to the collection.