Default transformers no longer duplicate on each request
When you had a default transformer (e.g. transformPath) configured on
resource-api, it was duplicated for each and every request that was
made, so it would be applied once on the first request, twice on the
second request, etc. The end result is that having default transformers
would almost certainly break multiple requests.

This was caused by a misuse of angular.extend when initializing configs.

Since angular.extend() performs a shallow copy, its use to extend the
intential constant `emptyConfig` actually modified the array values so
they were unintentionally populated for future requests.