Resources are considered reloading on initial load
While the original request to load data for a resource or collection is
still pending, any call to reload() for that resource will return the
original promise. This means you can safely call reload() on a
just-returned resource to ensure that it is reloaded if need be without
worrying whether it will fire of an unnecessary http request.