Do not duplicate config transformers on reload
Whenever a reload took place, through api.reload() or $reload(), any
transformers (request, response, or path) were duplicated for
subsequent reloads. This bug was similar to a previous one that affected
all requests and involved only default transformers, but it differed in
that it was only now affecting reload requests and was not limited to
default transformers only.