api.put() sends original $proxy values in http request
If you give a resource object with proxied properties to api.put(), it
should now send the original value from when the property was configured
to be proxied in the http request rather than the result of the
proxying. The proxying functionality is generally used for lazy loading
other resource objects, which is rarely a behavior that apis consider.
Plus, if you use proxying enough, you'll likely find yourself in a
situation where there is too much recursion for the json conversion to
work.