[#7] Fix bug where api.reload() choked on certian resources
The api.reload function, when given a resource that had an array on the
`resources` property, would trigger a JS error as it made the incorrect
assumption that it was trying to reload a collection rather than a
resource.

This fix is really just a bandaid rather than a complete fix.